pyproject.toml
[project]
name = "modeling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "omegaconf>=2.3.0",
    "pandas>=2.3.3",
    "ruff>=0.14.4",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
]

# Ruff Format Settings   ---- should be fixed with bluejun0901.
[tool.ruff]
# Exclude files
exclude = [
    ".venv",
    "data",
    "notebooks",
    "__pycache__"
]

line-length = 120      # Maximum chars in one line

# Ruff Linter
[tool.ruff.lint]
# Rules : 
# * E, W : Error, Warnings
# * F : Logical Error
# * I : Import Sort
select = ["E", "W", "F", "I"]
ignore = []

fixable = ["ALL"]
unfixable = []


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"


[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401", "F403"]
"**/**/__init__.py" = ["F401", "F403"]
